pyproject.toml
[tool.poetry]
name = "magic-sigexplorer"
version = "0.1.0"
description = ""
authors = ["The Haag <5632822+MHaggis@users.noreply.github.com>"]
license = "Apache 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
PyYAML = "^6.0.1"
Jinja2 = "^3.1.2"
mkdocs = "^1.5.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
